Update patch set 3
Patch Set 3:

In the future, we need to re-introduce the notion of the OEM error codes by providing a separate field for them, maybe an extra opaque U32 value or something, but INTERNAL_ERROR vs MODEM_ERROR is not specific enough to even do meaningful statistics collection. In addition, there are specific meanings associated with INTERNAL_ERROR,  MODEM_ERROR that imply implementation details we cannot/should not imply. The naming is fantastic too... Internal error is an error generated on-modem (for us modem == radio). System error could literally be anything, including that a function is unsupported, the radio isn't present, or that it's the wrong day of the week. It's so broad that we should never use it (imo, lmk). And the beautifully named MODEM_ERROR is actually an error generated by the Vendor RIL, which implies that there is a RIL, when there may or may not be a RIL... that's pure implementation detail. I picked INTERNAL_ERROR because even though the name is bizarre, it basically says "radio hit an unexpected scenario", which is at least constrained enough to mean that the implementers didn't expect this to happen in normal operation (meaning it can't imply unsupported), and it doesn't imply that a RIL exists.

Quoted for reference:
INTERNAL_ERR - Modem hit unexpected error scenario
SYSTEM_ERR - Hit platform or system error
MODEM_ERR - Vendor RIL got unexpected or incorrect response from modem

Patch-set: 3
Reviewer: Gerrit User 1060151 <1060151@85c56323-6fa9-3386-8a01-6480fb634889>