Fix LTE Timing Advance Doc String
The LTE TA Doc string has been fixed to reflect
the actual reported value of the field. The spec
reference was previously incorrect, which would
have indicated a range of 0-63 as the incremental
timing advance; however the field actually reports
the absolute timing advance, which has a much
larger range of 0-1282. In addition, this CL
updates the comment to remove a link to a defunct
website and explicitly states the range extracted
from the updated spec reference. This does not
impact the current behavior of the API as the the
previously incorrect docstring was nonsensical.

Bug: 66751464
Test: compilation (doc only change)
Change-Id: I2fc90c08ed6dd14548b10c638bf805b37d5c34e8