Update patch set 3
Patch Set 3:

I think the current patch has proposed both some items in keeping with consistency of other HALs and one change that is a philosophical tightening of the API... we can avoid making any philosophical changes without allowing things like EINVALIDARGUMENT when there is a method that takes no arguments... in 1.0/1.1 if we did that it was just a mistake. I don't want to keep this thread on life support if we're not going to take it, but in terms of the magnitude of change, the current HAL has at least one method that's significantly more lenient than equivalent methods in 1.0/1.1. Knowingly not fixing that is a philosophical loosening of the HAL contract.

Patch-set: 3
Tag: mailMessageId=<CADhJOfZrX=rQhrP85kVJUSysWoaaN14VtVO6qZwBj0oBfoinrA@mail.gmail.com>