Update patch set 1
Patch Set 1:

This is good for the specific thing here, but man oh man, I was having some shower thoughts. I believe we should significantly expand some of the documentation around these indications. The HAL documentation is the public documentation for all of these and I can see how easy it would be for vendors to be confused unless they have access to our internal design docs. We should distill and publish some of that.

Patch-set: 1
Reviewer: Gerrit User 1060151 <1060151@85c56323-6fa9-3386-8a01-6480fb634889>