Update patch set 1
Patch Set 1:

Not sure whether this is needed: the test module name is wrong, but there's no VTS test project for EVS that uses this name.

About the CP on mr1: even if this CP is needed, the mr1 CP could also be merged because otherwise there will be a resolve merge conflict CL required in mr1.

Patch-set: 1
Reviewer: Gerrit User 1132849 <1132849@85c56323-6fa9-3386-8a01-6480fb634889>