Add a Test for RadioBandMode to Ensure Auto Mode
The radio band mode is perilous because depending
on the setting and the detected locale, other
settings may be disabled. That can leave the modem
in a soft-brick state. Thus, BandMode = 0 = AUTOMATIC
should always be supported so that the modem can
update and select a band mode of its choosing.

Bug: 28124606
Test: vts radio - getAvailableBandModes
Change-Id: I1f25d9125ef06d290e3a89b5d2162c6bfe939eba