Mark configstore-utils as double_loadable
configstore-utils is explicitly marked as double_loadable since it is one of the
(indirect) dependencies of the LLNDK library libvulkan
and at the same time the lib itself is marked as VNDK. Such lib can be
double loaded inside a vendor process.

Note: even without this change, the library is already capable of being
double loaded due to the dependency graph around it. This change is to
make it explicit so that double loading of a library is carefully
tracked and signed-off by the owner of the lib.

Bug: 77155589
Test: m -j
Merged-In: Ibece0476b819e3387836305ee3b1c9dbf614920b
Change-Id: Ibece0476b819e3387836305ee3b1c9dbf614920b
(cherry picked from commit 9fb3c20bc62a3b2977cf2028cfa1c8c4938ee5b2)