Update patch set 2
Patch Set 2:

We use the worst case to update timeout to 5 minutes.
Total time is 70s + 120s + 88s for all rat because we search PLMN in global mode.

The worst case for GSM is about 70 seconds.
For GSM, normal search time in GSM would be about 24 seconds.
The worst case would be in presence of high interference signal, where GSM could detect signal strength but cannot decode cell information.
GSM tries 100 cells during the search procedure and this could take about 70 seconds with GSM re-attempt to decode these bad blocks in poor radio condition.

The worst case for 3G is about 120 seconds.
3G part is dependent on following:
1. UE supported bands and bands supported in that country
2. Number of PLMN’s available and its cell signal strength
3. Along with PLMN list, if CTCH reception is going on
4. During PLMN list, PLMN list might be suspended and then restarted for dual SIM case.
Considering UE supports 5 bands 1, 2, 4, 5, and 8.
1. If signal strength of the found cells are good, then 3G PLMN list will be completed within 45 seconds.
2. If signal strength is not good or CTCH reception is configured, then it will take long time. As per our observation, it didn’t exceed 2 minutes.
3. For dual SIM case, it depends on dual SIM activity.

The worst case for LTE is about 88 seconds.
For LTE part, it depends on UE supported bands and signal status in environment.
The worst case is to search on all bands and there are strong signal on whole band. For example, our modem can support 18 bands. They are [1],[2],[3],[4],[5],[8],[12],[13],[14],[17],[25],[26],[30],[38],[40],[41],[66],[71]. Total bandwidth is 928MHz and the time of cell search on 928MHz is about 80 seconds.
And we also need to consider some overhead (e.g. message exchange or search algorithm or time of SIB receiving), so total time of the worst case is about 88 seconds.

Patch-set: 2
Reviewer: Gerrit User 1226402 <1226402@85c56323-6fa9-3386-8a01-6480fb634889>