Update patch set 1
Patch Set 1:

(5 comments)

Thanks for your feedback.
The architecture of this interface assumes that we support a single person face enrollment per device user. We have designed and implemented the entire face authentication stack (API and framework) under this agreement. I know you were considering to change this back to multiple enrollments, but in the last sync you mentioned we should keep the single enrollment design. Would you like to reconsider and we will modify the entire stack and the HAL accordingly? We would prefer to do this big change once you finalize the decision on multiple enrollments. Thanks.

Patch-set: 1
CC: Gerrit User 1284969 <1284969@85c56323-6fa9-3386-8a01-6480fb634889>