Allow general errors for getImsiForApp().
This is to match it with other 1.0 tests where general errors are
allowed. For newer tests we have decided to not allow these errors,
but a failure for this old test is reported when run with SIM
present.

Test: run vts -m VtsHalRadioV1_0Target
Bug: 109889468
Change-Id: If36083b7832706a50805932e8ba08e4eb397f3fe
Merged-In: If36083b7832706a50805932e8ba08e4eb397f3fe
(cherry picked from commit 817848e59e18334a30f623ef143cf182716fd3fa)