Allow RadioError::NONE for startLceService
Allow startLceService to succeed even in the
SIM_ABSENT case. The original RIL documentation
states that SUCCESS is valid for all LCE
operations, and there is no logical reason why
one of these operations must fail in a no-SIM
case (though a vendor may choose not to support
that configuration). Thus, a call to
startLceService should permit NONE when requested
in the no-SIM case for 1.0 VTS.

Bug: 110181475
Test: compilation - verification pending a 1.0
    device that returns NONE in this case.

Change-Id: Ib47f723b5454eb415234a864445ad97f0b195b67