Update patch set 1
Patch Set 1:

I don't fully understand the problem here. Yes, including the header file under the namespace is a little weird. But, we had to do that since the old HAL header file is a plain "C" header file and it was being used by a bunch of other projects. So, we put in this hack to move all the header contents into a namespace in our project without messing with the original header.


Is this wifi_legacy_hal.h file being included by other projects? What is this breaking?

Patch-set: 1
Reviewer: Gerrit User 1074459 <1074459@85c56323-6fa9-3386-8a01-6480fb634889>