Audio HAL: Use a cache for devices
openDevice(name) was always creating a new device, even if the
same name was provided.
Most HAL do not support to have the same IDevice open twice,
thus leading to crashes/bugs if it occurred.

Even if the client is careful no to open the same device twice,
it can still occur if the client crashes, or drops the sp to the
device and repoen it at destruction is not synchronous.

Thus implement a cache (similar to dlopen) of devices.
When a device that was already open is reopen, return the same
object.

The cache references the IDevice s via week pointers thus does not
impact the IDevicei s life cycle.

Test: flash and check there is audio
      atest VtsHalAudioV2_0Target
Change-Id: I175c4e30702e312e22eba62eebd6d82854bf0b71
Signed-off-by: Kevin Rocard <krocard@google.com>