Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > (3 comments)
> > 
> > Presumably, we could just call setDataProfile() in a loop when a new sub is inserted. Then the setAvailableDataProfiles() would just take an array of ints:
> > 
> > setAvailableDataProfiles(int[] profileIds);
> 
> setDataProfile is already used for other purpose. It's called before setup data call and initial data attach. Also profile id is a VZW specific thing.

We don't have anything VZW-specific in AOSP. We either have bugs or the modems have bugs. We couldn't test a carrier-specific feature in VTS, so it doesn't exist. :-)

I'm aware of what we're using setDataProfile to do now, but there's no conceptual reason why we shouldn't be able to use it more generally. We still need a new API to meet our current needs, but if we can separate APN config updating from APN availability updating I think that would be preferable to sending down a whole pile every time. To wit, if we are able to identify these profiles by ID and the carrier/user updates a profile, we don't want to hook APN updating logic into this mechanism to re-send the Available APNs; that would be janky.

I'm also looking very skeptically at profileId... that seems like a reasonable concept, and I believe we're seeing a case where it should be used. @Niraj, how badly would things break if we started using profileId across the board as a unique key for identifying APN DB entries in communications with the modem?

Patch-set: 1
Reviewer: Gerrit User 1060151 <1060151@85c56323-6fa9-3386-8a01-6480fb634889>