Note confusion-causing breaks in HIDL convention.
These were either made at a time when the convention was not fully
formed or missed during review.

It is somewhat misleading since method overloading isn't supported
and method names need this kind of prefix, but nothing else does.

The reason for this is that everything is namespaced, but methods are
all in the same namespace. The reason method overloading is not supported
in HIDL is because the HIDL types may map to types that collide in the
target languages, and this would cause any sort of overloading rules
to require complicated machinery.

Bug: N/A
Test: hidl-gen -Lcheck android.hardware.usb@1.1 android.hardware.vibrator@1.1 && echo ":)"
:)

Change-Id: Iac23c9311925ed140ff1e15d1366829b078c8866