Update patch set 3
Patch Set 3:

> is this need in pie? if not how about merging to master first?

It was hit a couple of times by OEM. Their HAL would crash and then the test would also crash, resulting two crash backtrace.
That makes it confusing for the OEM as it thinks the test crash is the issue instead of the HAL crash.

Additionally the fix is trivial.

Patch-set: 3
Reviewer: Gerrit User 1136369 <1136369@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0