[WIP] [DO NOT SUBMIT] NNAPI Burst prototype -- HAL
This is a prototyping effort.

FastMessageQueue is a Treble-compliant data structure that enables fast
communication between two processes. The FMQ object itself is an atomic
circular buffer that is optionally synchronized with a futex. However,
FMQ has no notion of ownership or lifetime across processes, so it must
be paired with higher-level constructs to manage the lifetime and
ownership.

This topic introduces the notion of a "ExecutionBurstChannel", which is
the FMQ object paired with a HIDL interface object for ownership and
ref-counting purposes. It is a unidirectional flow of information with
exactly one producer and exactly one consumer. In this prototype, the
object is always owned by the sender and referenced by the receiver.
A complete sequence of ExecutionBurstChannels compose an "Execution
Burst", or more simply "Burst".

This CL includes three different changes:
1) Defines the ExecutionBurstChannel structure in IFmqContextToken.hal
   and adds the setup method IPreparedModel::configureExecutionBurst in
   IPreparedModel.hal.
2) Tests the Burst in both the execution path and validation path in the
   Vendor Test Suite (VTS) in neuralnetworks/1.*/vts/functional/.
3) Creates a simple native Google Benchmark test to check the
   performance of three HAL models (add_float, mobilenet_quantized, and
   mobilenet_float) when running through three execution paths (original
   execution, blocking FMQ with futex, and spinning FMQ).

The VTS binary--VtsHalNeuralnetworksV1_2TargetTest--can be built and run
as any previous version could.

The two new native benchmarks can be built using:
  m BurstBenchmark -j40
  m BurstFunctional -j40
Pushed to the device using:
  adb push $OUT/data/benchmarktest64/BurstBenchmark/BurstBenchmark /data/local/tmp/BurstBenchmark
  adb push $OUT/data/nativetest64/BurstFunctional/BurstFunctional /data/local/tmp/BurstFunctional
And run using:
  adb shell /data/local/tmp/BurstBenchmark [--benchmark_repetitions=N]
  adb shell /data/local/tmp/BurstFunctional [--gtest_filter=*]

The only local modification to this CL is to change the name of the
service being tested in BurstBenchmark and BurstFunctional. Currently
it's pre-set to "sample-all" in BurstBenchmark line 76 and
BurstFunctional line 97.

Bug: 119570067
Test: mma
Test: BurstBenchmark
Test: BurstFunctional
Test: VtsHalNeuralnetworksV1_2TargetTest
Change-Id: Ia7136151289e5f9e04486718702502016ee72f67
(cherry picked from commit 8c2e706567ab60077b6829c84de09e5d0f12cf17)