Update patch set 1
Patch Set 1:

Of course, everything about this CL is subject to change. It is being moved into AOSP for based on this reasoning (copied from the bug):

So, this is the process we've been following (see b/68112778 for instance):
- have (compatibility_matrix.)current.xml at 'level #' (corresponding to the next dessert release) only in master
- when the dessert releases, it also has current.xml
- at integration complete for the dessert, it gets renamed to #.xml and once that merges into master, a copy of it is made for level #+1 called which is current.xml.

This is okay since in practice the only device that is at the next dessert release's FCM level is the next Pixel device which is not on AOSP. However, this causes two problems (maybe more):
- devs (like nharold) can't add hals being developed in AOSP to the compatibility matrix in AOSP
- when teams want to deprecate an interface from the next release, they have to do it in AOSP. This is much less visible, and it would be nicer for partners if this change could be made directly in AOSP.

So, this is the process that we'll change to:
- have current.xml at 'level #' corresponding to the next dessert release in AOSP->master.
- when the next dessert releases, it also has current.xml
- at integration complete for the dessert, we make a copy of current.xml and call it #.xml and then update the level in current.xml to be '#+1'.

This actually requires one less CL to do which will be nice. The downside is that the release branch will have a current.xml file. We can fix this by adding a new variable which prevents current.xml from being used in the release branch (where it is meaningless at level #+1). We could also delete current.xml and restore it in the $DT-plus-aosp, but I think that makes it even more complicated. I'll give this more thought and move things over soon.

Patch-set: 1
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>