Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> From the internal review:
> > It is not a change in behavior since the HAL implementation determines the data formulation/passing.
> 
> Please confirm that this will not be accompanied by a change in behavior (or that there was not a previous change in behavior) and that this documentation reflects the actual original behavior of implementations.

There actually was a recent change in the reference HAL implementation that is accompanied by this change. The purpose of this change is to guide OEMs to implement it that way, too, since it is stated by the Android docs on sensors.

Is there a more appropriate way to change HAL documentation in this type of scenario?

Patch-set: 1
Reviewer: Gerrit User 1175853 <1175853@85c56323-6fa9-3386-8a01-6480fb634889>