Update patch set 8
Patch Set 8:

(6 comments)

Hey Paul, btw, it looks like this is conflicting with https://android-review.googlesource.com/c/platform/hardware/interfaces/+/812974/1.

The interface lgtm. There are a couple of legacy things that would be nice to be fixed. It looks like that might be done on the other CL though, so perhaps you can just rebase on top of that.

Patch-set: 8
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>