VTS: add 20ms delay bfore checking the event count
When the full VTS test suite is run,the test BluetoothHidlTest.
LoopbackModeBandwidth_32bit fails, but passes when run separately.
We could see that this is caused due to race condition between main thread
and binder thread of Bluetooth VTS process. The event_queue is getting
updated in another thread after the validation is done from main thread.
Test may fail sometimes due to this race condition

Bug: 120460769

Change-Id: If9190f8071737cf82cec566ab3bc98afe68fde24