Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-2
> 
> we should never fix race conditions in tests with sleeping

Agree this is not good fix and the issue deserves a better one. However this is not a really "condition race" issue, the test case is now waiting on the first event and try to read all the following events when the first event is received, it should be fixed as waiting on events until the all the expected events are read out or an timeout happens.

Patch-set: 2
CC: Gerrit User 1207838 <1207838@85c56323-6fa9-3386-8a01-6480fb634889>