Audio HAL: Do not release resource in destructor
HIDL destructor can be infinitely delayed as their invocation
is dependant on proxy binder destruction detection which is
concurrent to other binder commands.

Thus as the destructor invocation can not be reliably sequenced with
other method call, it must not do anything.

As a result move all its code in the predictable and sequencable close()
method.

Test:  atest VtsHalAudioV4_0TargetTest
Bug: 118655804
Change-Id: I1a5ec28bce9802ec654c139153ec4aa6786474e5
Signed-off-by: Kevin Rocard <krocard@google.com>