Update patch set 1
Patch Set 1:

> Please see my comments on the bug. This might cause re-introduction of the issues described in b/34800063 and b/34499806. They were the reason we have moved thread joining and device closing into the destructor.

Ok so this is going to be a non trivial fix. Either we need to make the thread join faster (the fast message queue wait should have a termination signal) (proper fix)
Or we need to make the join and close_output_stream async and wait for it to finish when open_stream is called (workaround).

I do not have the time to work on the issue right now. So I will add a usleep in the test after the close and destroy.

Patch-set: 1
Reviewer: Gerrit User 1136369 <1136369@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0