Update patch set 1
Patch Set 1:

Benjamin, is this the topic of cherry-picks?

They look good, but you'll want to add Merged-In (and for future reference, preferably cherry-pick with -x so that the sha from the old commits is in the message as well). Otherwise, they usually won't pass TH, and they can cause weird symptoms. Otherwise, lg!

Patch-set: 1
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>