Update patch set 2
Patch Set 2:

The new patchset  adds a sleep after the close as the current design is racy. I do not see any vts test results with the CL on  pixel devices. Please do attach it to the CL for review. We want to avoid regressions to this branch.

Patch-set: 2
Reviewer: Gerrit User 1153368 <1153368@85c56323-6fa9-3386-8a01-6480fb634889>