Update patch set 9
Patch Set 9:

(5 comments)

Further testing on boot patch level and the like will have to wait until we consult with the treble team on the best way to handle it, potentially through libfs_avb. I recommend that we take this CL now and get it checked in, and then follow up with proper patchlevel checking according to the treble team afterwards.

Patch-set: 9
Reviewer: Gerrit User 1115498 <1115498@85c56323-6fa9-3386-8a01-6480fb634889>