Update patch set 1
Patch Set 1:

Can you merge in the Q release branch first? It's important that the final versions of interfaces are there. Otherwise, there will exist a state where there is an interface hash in the Q branch but not in master (and this will cause master VTS to fail on Q devices), and so it could result in failure bugs (or more failures next year if this is missed in Q).

Patch-set: 1
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>