Revert "keymaster 3.0: make service use nobody as user and remove system group"
This reverts commit 848607a121d20d61f76bfe81275038f6f18c49e8.

Reason for revert: b/132758952
We found keymaster3 is not allowed to access attestation keybox because keymaster3 service user is set to "nobody" and group is set to drmrpc only. But we injected keybox with an HIDL service "keyattestation" with user/group set to "system". Therefore, we encountered an issue on EC/RSA key attestation GTS test case.

Change-Id: I7301e5d9f9c88c87bc18b6f0e4542e79bb09b2a6