Update patch set 2
Patch Set 2:

Overall the HAL interface looks good.  
Internally the next steps is to get this prioritized so it can be scheduled.  
To Actually be added as a new HAL it needs VTS tests and a acould (cuttlefish) implementation.

And all of that needs formal review.

Patch-set: 2
Reviewer: Gerrit User 1060286 <1060286@85c56323-6fa9-3386-8a01-6480fb634889>