gralloc: add get/set BufferMetadata
Add get and set functions that allow buffer metadata to be set
on the buffer.

This patch is for public vendor/OEM review. We are interested in
learning if the BufferMetadata proposed in the patch is sufficients.
Is there additional metadata that we should add to the api? Should
the proposed metadata be changed in some way?

If you have any thoughts, ideas or concerns, please leave a comment.

This api uses a proposal from ARM to describe plane layout. The
plane layout proposal is uploaded by ARM as a seperate patch.

Bug: 138457127
Test: N/A

Change-Id: Ia9544f75703453cb2f39c945d6c0a4966151c8c5