gralloc: add get and set BufferMetadata support
Add get and set functions that allow buffer metadata to be set
on the buffer.

This patch is for public vendor/OEM review. We are interested in
learning if the BufferMetadata proposed in the patch is sufficients.
Is there additional metadata that we should add to the api? Should
the proposed metadata be changed in some way?

If you have any thoughts, ideas or concerns, please leave a comment.

This patch is a combination of the following patches. They have been
squashed (and updated) to make them easier to review. Two of the patches
are proposals by ARM that have been incorporated into this patch.

https://android-review.googlesource.com/c/platform/hardware/interfaces/+/1107574
https://android-review.googlesource.com/c/platform/hardware/interfaces/+/1109946
https://android-review.googlesource.com/c/platform/hardware/interfaces/+/1124857
https://android-review.googlesource.com/c/platform/hardware/interfaces/+/1126861

Bug: 141632767, 141631993, 141631415
Test: N/A
Change-Id: I0e5dc43d8de43104b38905da15b6f3636851404d