Update patch set 3
Patch Set 3:

(4 comments)

Thanks for the comments Naseer! 

Just some background on the patches, we are currently waiting for ARM to push a patch where they have defined per plane metadata. Once they are able to push the patch, we have a couple of ideas for additional per plane metadata that we would to add.

Once everything is in place, I will send an email to everyone to let them know that the patches are up for review.

Patch-set: 3
Reviewer: Gerrit User 1073546 <1073546@85c56323-6fa9-3386-8a01-6480fb634889>