Add Valid maxSearchTime to startNetworkScan 1_2 + 1_4
The maxSearchTime field is documented to be 60-3600; however
the startNetworkScan basic test doesn't initialize this field,
which means it has a value of zero. This CL updates that field
to use the minimum documented time of 60.

In addition, this CL updates the incrementalResults/Periodicity
fields to be explicitly initialized, and sets the periodicity to
1, which will improve compatibility on older HALs that relied on
an exception clause to exempt periodicity from range-checking when
incremental reporting was disabled.

Bug: 140451434
Test: make vts && vts-tradefed run commandAndExit vts -m \
      VtsHalRadioV1_2Target -t RadioHidlTest_v1_2#startNetworkScan
Change-Id: I302536e9743ffd25f525b9c8d1ee59ccb1b42d11